requirements.txt
python-telegram-bot==13.15
playwright>=1.49.0
apscheduler~=3.6.3
requests>=2.31.0

python-dotenv
beautifulsoup4>=4.12.0
lxml>=4.9.0
tzdata
